Included main.cpp, parser.cpp and its header, and wirings.txt
The main.cpp file includes all the functions regardind the ciphering logic of the Enigma.
The parser.cpp and parser.h files includes the functions for reading wirings.txt file.
A .txt file is used for configuration purposes so the project can be easily updated in the near future, including more types of rotors, for example.